added obsseq test program that includes obs_space_diagnostics
test_obsseq_rw_space includes a call to obs_space_diagnostics that is copied into the file.
test_obsseq_rw_space also includes a call to filter_setup_obs_sequence which was made public inside filter.mod.
This was last tested with gfortran and was able to compile and run.